Revert "Streaming exports (#1826)"
This reverts commit 61d2940a5ff53a755240d661f94d430eecc8b543.
